Refactor app.rb and create create_instances_module.rb
